// src/clients/CatalogAdministration/models/SearchRedirect.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kibo Catalog Admin Services
 * OpenAPI Spec for Kibo Catalog Admin Services
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { AuditInfo } from './AuditInfo';
import {
    AuditInfoFromJSON,
    AuditInfoFromJSONTyped,
    AuditInfoToJSON,
} from './AuditInfo';

/**
 * Redirects based on what is entered in site search
 * @export
 * @interface SearchRedirect
 */
export interface SearchRedirect {
    /**
     * 
     * @type {string}
     * @memberof SearchRedirect
     */
    redirectId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SearchRedirect
     */
    url?: string | null;
    /**
     * 
     * @type {number}
     * @memberof SearchRedirect
     */
    statusId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof SearchRedirect
     */
    tenantId?: number;
    /**
     * 
     * @type {number}
     * @memberof SearchRedirect
     */
    siteId?: number | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof SearchRedirect
     */
    searchTermList?: Array<string> | null;
    /**
     * 
     * @type {AuditInfo}
     * @memberof SearchRedirect
     */
    auditInfo?: AuditInfo;
}

/**
 * Check if a given object implements the SearchRedirect interface.
 */
export function instanceOfSearchRedirect(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function SearchRedirectFromJSON(json: any): SearchRedirect {
    return SearchRedirectFromJSONTyped(json, false);
}

export function SearchRedirectFromJSONTyped(json: any, ignoreDiscriminator: boolean): SearchRedirect {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'redirectId': !exists(json, 'redirectId') ? undefined : json['redirectId'],
        'url': !exists(json, 'url') ? undefined : json['url'],
        'statusId': !exists(json, 'statusId') ? undefined : json['statusId'],
        'tenantId': !exists(json, 'tenantId') ? undefined : json['tenantId'],
        'siteId': !exists(json, 'siteId') ? undefined : json['siteId'],
        'searchTermList': !exists(json, 'searchTermList') ? undefined : json['searchTermList'],
        'auditInfo': !exists(json, 'auditInfo') ? undefined : AuditInfoFromJSON(json['auditInfo']),
    };
}

export function SearchRedirectToJSON(value?: SearchRedirect | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'redirectId': value.redirectId,
        'url': value.url,
        'statusId': value.statusId,
        'tenantId': value.tenantId,
        'siteId': value.siteId,
        'searchTermList': value.searchTermList,
        'auditInfo': AuditInfoToJSON(value.auditInfo),
    };
}